Move images into component which uses them
