+++ src/components/MMBoard/images/cards/card-acceptance.png
