+++ src/components/MMBoard/images/cards/card-curiosity.png
