+++ src/components/MMBoard/images/cards/card-goal.png
